added single layer simulation
