+++ maria/atmosphere/__init__.py
@@ -19,7 +19,6 @@ here, this_filename = os.path.split(__file__)
 
 from .. import base
 
-
 class BaseAtmosphericSimulation(base.BaseSimulation):
     """
     The base class for modeling atmospheric fluctuations.
@@ -67,7 +66,7 @@ class BaseAtmosphericSimulation(base.BaseSimulation):
                                                                                  self.site.spectrum.side_elevation),
                                                                                 (self.site.spectrum.trj * passband).sum(axis=-1))
 
-                self.data[band_mask] = band_T_RJ_interpolator((self.integrated_water_vapor[band_mask], np.degrees(self.EL[band_mask])))
+                self.data[band_mask] = band_T_RJ_interpolator((self.effective_integrated_water_vapor[band_mask], np.degrees(self.EL[band_mask])))
 
         if units == 'F_RJ': # Fahrenheit Rayleigh-Jeans 🇺🇸🇺🇸🇺🇸
 
@@ -77,610 +76,236 @@ class BaseAtmosphericSimulation(base.BaseSimulation):
 
                     
 
-DEFAULT_LA_CONFIG = {
+DEFAULT_ATMOSPHERE_CONFIG = {
     "min_depth": 500,
     "max_depth": 3000,
     "n_layers": 4,
     "min_beam_res": 4,
 }
 
-class LinearAngularSimulation(BaseAtmosphericSimulation):
-    """
-    The linear angular model treats the atmosphere as a bunch of layers. 
 
+
+def get_rotation_matrix_from_skew(x):
+    S = x * np.array([[0, 1], [-1, 0]])
+    R = sp.linalg.expm(S)
+    return R
+            
+MIN_SAMPLES_PER_RIBBON = 2
+JITTER_LEVEL = 1e-6
+
+class SingleLayerSimulation(BaseAtmosphericSimulation):
+    """
+    The simplest possible atmospheric model.
     This model is only appropriate for single instruments, i.e. when the baseline is zero.
     """
 
-    def __init__(self, array, pointing, site, config=DEFAULT_LA_CONFIG, verbose=False, **kwargs):
+    def __init__(self, array, pointing, site, layer_height=1e3, min_beam_res=8, verbose=False, **kwargs):
         super().__init__(array, pointing, site, **kwargs)
 
-        self.config = config
-        for key, val in config.items():
-            setattr(self, key, val)
-            if verbose:
-                print(f"set {key} to {val}")
+        self.min_beam_res = min_beam_res
 
-        self.layer_boundaries = np.linspace(self.min_depth, self.max_depth, self.n_layers + 1)
+        self.layer_height = layer_height
 
-        # layer of integrated atmosphere
-        self.layer_depths = 0.5 * (self.layer_boundaries[1:] + self.layer_boundaries[:-1])
-        self.layer_thicks = np.diff(self.layer_boundaries)
+        # this is approximately correct
+        self.layer_depth = self.layer_height / np.sin(np.mean(self.pointing.el))
 
-        # returns a beam waists and angular waists for each frequency for each layer depth
-        self.waists = self.array.get_beam_waist(
-            self.layer_depths[:, None], self.array.primary_size, self.array.dets.band_center.values
+        # returns the beam waist for each detector at the layer distance
+        self.physical_beam_waists = self.array.get_beam_waist(
+            self.layer_height, self.array.primary_size, self.array.dets.band_center.values
         )
-        self.angular_waists = self.waists / self.layer_depths[:, None]
+        self.angular_beam_waists = self.physical_beam_waists / self.layer_depth
 
-        self.min_ang_res = self.angular_waists / self.min_beam_res
+        self.ang_res = self.angular_beam_waists.min() / self.min_beam_res
 
-        self.weather.generate(time=self.pointing.time, fixed_quantiles=self.site.weather_quantiles)
+        self.weather = weathergen.Weather(region=self.site.region, 
+                                          altitude=self.site.altitude,
+                                          diurnal=True, 
+                                          seasonal=True,
+                                          levels=np.array([self.site.altitude + self.layer_height]))
 
-        self.heights = self.site.altitude + self.layer_depths[:, None] * np.sin(self.pointing.el)[None, :]
+        
 
-        for attr in [
-            "water_vapor",
-            "temperature",
-            "pressure",
-            "wind_north",
-            "wind_east",
-        ]:
+        self.weather.generate(time=self.pointing.time.mean(), 
+                              mode="median", 
+                              fixed_quantiles=self.site.weather_quantiles)
 
-            setattr(
-                self,
-                attr,
-                sp.interpolate.RegularGridInterpolator(
-                    (self.weather.levels, self.weather.time),
-                    getattr(self.weather, attr),
-                )((self.heights, self.pointing.time[None])),
-            )
+        angular_velocity_x = (+self.weather.wind_east[0] * np.cos(self.pointing.az) 
+                                   -self.weather.wind_north[0] * np.sin(self.pointing.az)) / self.layer_depth
+        
+        angular_velocity_y = (-self.weather.wind_east[0] * np.sin(self.pointing.az) 
+                                   +self.weather.wind_north[0] * np.cos(self.pointing.az)) / self.layer_depth
 
-            # setattr(self, attr, sp.interpolate.interp1d(self.weather.height, getattr(self.weather, attr), axis=0)(self.heights))
-
-        self.wind_bearing = np.arctan2(self.wind_east, self.wind_north)
-        self.wind_speed = np.sqrt(np.square(self.wind_east) + np.square(self.wind_north))
-
-        self.layer_scaling = self.water_vapor * self.temperature * self.wind_speed * self.layer_thicks[:, None]
-        self.layer_scaling = np.sqrt(np.square(self.layer_scaling) / np.square(self.layer_scaling).sum(axis=0))
-
-        # the velocity of the atmosphere with respect to the array pointing
-        self.AWV_X = (
-            +self.wind_east * np.cos(self.pointing.az[None]) - self.wind_north * np.sin(self.pointing.az[None])
-        ) / self.layer_depths[:, None]
-        self.AWV_Y = (
-            (-self.wind_east * np.sin(self.pointing.az[None]) + self.wind_north * np.cos(self.pointing.az[None]))
-            / self.layer_depths[:, None]
-            * np.sin(self.pointing.el[None])
-        )
-
-
-        center_lon, center_lat = utils.get_center_lonlat(self.pointing.az, self.pointing.el)
-        self.pointing.dx, self.pointing.dy = utils.lonlat_to_xy(self.pointing.az, self.pointing.el, center_lon, center_lat)
+        print(angular_velocity_x.shape)
+        print(angular_velocity_y.shape)
+        
+        # compute the offset with respect to the center of the scan
+        center_az, center_el = utils.get_center_lonlat(self.pointing.az, self.pointing.el)
+        self.pointing.dx, self.pointing.dy = utils.lonlat_to_xy(self.pointing.az, self.pointing.el, center_az, center_el)
 
         # the angular position of each detector over time WRT the atmosphere
-        self.REL_X = (
-            self.array.sky_x[None, :, None]
-            + self.pointing.dx[None, None]
-            + np.cumsum(self.AWV_X * self.pointing.dt, axis=-1)[:, None]
-        )
-        self.REL_Y = (
-            self.array.sky_y[None, :, None]
-            + self.pointing.dy[None, None]
-            + np.cumsum(self.AWV_Y * self.pointing.dt, axis=-1)[:, None]
-        )
+        # this has dimensions (det index, time index)
+        self.boresight_angular_position = np.c_[self.pointing.dx + np.cumsum(angular_velocity_x * self.pointing.dt, axis=-1),
+                                                self.pointing.dy + np.cumsum(angular_velocity_y * self.pointing.dt, axis=-1)]
 
-        self.MEAN_REL_X, self.MEAN_REL_Y = self.REL_X.mean(axis=1), self.REL_Y.mean(axis=1)
+        # find the detector offsets which form a convex hull
+        self.detector_offsets = np.c_[self.array.sky_x, self.array.sky_y]
+        doch = sp.spatial.ConvexHull(self.detector_offsets)
+        
+        # pad each vertex in the convex hull by an appropriate amount
+        unit_circle_angles = np.linspace(0, 2*np.pi, 32+1)[:-1]
+        unit_circle_offsets = np.c_[np.cos(unit_circle_angles), np.sin(unit_circle_angles)]
+        angular_padding_per_detector = 1.1 * self.angular_beam_waists
+        padded_doch_offsets = (angular_padding_per_detector[doch.vertices][:, None, None] * unit_circle_offsets[None, :] 
+                               + self.detector_offsets[doch.vertices][:, None]).reshape(-1, 2)
+        
+        # take the convex hull of those padded vertices
+        padded_doch = sp.spatial.ConvexHull(padded_doch_offsets)
+        detector_padding = padded_doch_offsets[padded_doch.vertices]
+        
+        
+        # add the padded hull to the moving boresight
+        atmosphere_offsets = (detector_padding[:, None] + self.boresight_angular_position[None]).reshape(-1, 2)
+        atmosphere_hull = sp.spatial.ConvexHull(atmosphere_offsets)
+        self.atmosphere_hull_points = atmosphere_offsets[atmosphere_hull.vertices]
 
-        # These are empty lists we need to fill with chunky parameters (they won't fit together!) for each layer.
-        self.para, self.orth, self.P, self.O, self.X, self.Y = [], [], [], [], [], []
-        self.n_para, self.n_orth, self.lay_ang_res, self.genz, self.AR_samples = (
-            [],
-            [],
-            [],
-            [],
-            [],
-        )
 
-        self.p = np.zeros((self.REL_X.shape))
-        self.o = np.zeros((self.REL_X.shape))
+        def optimize_area_minimizing_rotation_matrix(points):        
+            def ratio(x):
+                R = get_rotation_matrix_from_skew(x[0])
+                trans_points = R @ points[..., None]
+                log_ratio = np.log(trans_points[:, 0].ptp() / trans_points[:, 1].ptp())
+                #print(log_ratio)
+                return log_ratio
+            
+            res = sp.optimize.minimize(ratio, x0=[1.,], tol=1e-16, method="L-BFGS-B")
+            return res
+        
+        # R takes us from the real (dx, dy) to a more compact (cross_section, extrusion) frame
+        self.res = optimize_area_minimizing_rotation_matrix(self.atmosphere_hull_points)
 
-        self.layer_rotation_angles = []
-        self.outer_scale = 3e2
-        self.ang_outer_scale = self.outer_scale / self.layer_depths
+        #return
+        
+        assert self.res.success
+        self.R = get_rotation_matrix_from_skew(self.res.x)
 
-        self.theta_edge_z = []
+        #                   
+        #          ^      xxxxxxxxxxxx
+        #          |      xxxxxxxxxxxx
+        #   cross-section xxxxxxxxxxxx
+        #          |      xxxxxxxxxxxx
+        #          @      xxxxxxxxxxxx
+        #
+        
+        trans_points = self.atmosphere_hull_points @ self.R.T
+        cross_section_min = trans_points[:, 0].min()
+        cross_section_max = trans_points[:, 0].max()
+        extrusion_min = trans_points[:, 1].min()
+        extrusion_max = trans_points[:, 1].max()
+        
+        self.cross_section_side = np.arange(cross_section_min, cross_section_max + self.ang_res, self.ang_res)
+        self.extrusion_side = np.arange(extrusion_min, extrusion_max, self.ang_res)
+        
+        self.n_cross_section = len(self.cross_section_side)
+        self.n_extrusion = len(self.extrusion_side)
+        
+        CROSS_SECTION, EXTRUSION = np.meshgrid(self.cross_section_side, self.extrusion_side)
+        
+        self.TRANS_POINTS = np.concatenate([CROSS_SECTION[..., None], EXTRUSION[..., None]], axis=-1)
+        
+        extrusion_indices = [0, *(2 ** np.arange(0, np.log(self.n_extrusion) / np.log(2))).astype(int), self.n_extrusion-1]
+        
+        
+        
+        extrusion_sample_index = []
+        cross_section_sample_index = []
+        for i, extrusion_index in enumerate(extrusion_indices):
+        
+            n_ribbon_samples = np.minimum(np.maximum(int(self.n_cross_section * 2 ** -i), MIN_SAMPLES_PER_RIBBON), self.n_cross_section)
+            cross_section_indices = np.unique(np.linspace(0, self.n_cross_section - 1, n_ribbon_samples).astype(int))
+            cross_section_sample_index.extend(cross_section_indices)
+            extrusion_sample_index.extend(np.repeat(extrusion_index, len(cross_section_indices)))
 
-        radius_sample_prop = 1.5
-        beam_tol = 1e-2
+        self.extrusion_sample_index = np.array(extrusion_sample_index)
+        self.cross_section_sample_index = np.array(cross_section_sample_index)
+        
+        live_edge_positions = np.c_[self.cross_section_side, 
+                                    np.repeat(extrusion_min - self.ang_res, self.n_cross_section)]
+        
+        sample_positions = np.c_[CROSS_SECTION[self.extrusion_sample_index, self.cross_section_sample_index], 
+                                 EXTRUSION[self.extrusion_sample_index, self.cross_section_sample_index]]
+        
+        
+        # the sampling index will look something like:
+        #
+        #          x111010000000001 ...
+        #          x100000000000000 ...
+        #          x110000000000000 ...
+        #          x100000000000000 ...
+        #  leading x111000000000001 ...
+        #  edge    x100000000000000 ...
+        #          x110000000000000 ...
+        #          x100000000000000 ...
+        #          x111000000000001 ...
+        #
+        
+        n_live_edge = len(live_edge_positions)
+        n_sample = len(sample_positions)
+        
+        outer_scale = 0.1
+        
+        # sample upper {i,j}
+        i, j = np.triu_indices(n_sample, k=1)
+        COV_S_S = np.eye(n_sample) + JITTER_LEVEL
+        COV_S_S[i, j] = utils.functions.matern(np.sqrt(np.square(sample_positions[j] - sample_positions[i]).sum(axis=1)), outer_scale, 5/6)
+        COV_S_S[j, i] = COV_S_S[i, j]
+        
+        # this one is explicit
+        COV_LE_S = utils.functions.matern(np.sqrt(np.square(sample_positions[None] - live_edge_positions[:, None]).sum(axis=2)), outer_scale, 5/6)
+        
+        # live edge upper {i,j}
+        i, j = np.triu_indices(n_live_edge, k=1)
+        COV_LE_LE = np.eye(n_live_edge) + JITTER_LEVEL
+        COV_LE_LE[i, j] = utils.functions.matern(np.sqrt(np.square(live_edge_positions[j] - live_edge_positions[i]).sum(axis=1)), outer_scale, 5/6)
+        COV_LE_LE[j, i] = COV_LE_LE[i, j]
+        
+        inv_COV_S_S = utils.fast_psd_inverse(COV_S_S)
+        
+        self.A = COV_LE_S @ inv_COV_S_S
+        self.B = np.linalg.cholesky(COV_LE_LE - self.A @ COV_LE_S.T)
+        self.VALUES = np.zeros((self.n_extrusion, self.n_cross_section))
 
-        max_layer_beam_radii = 0.5 * self.angular_waists.max(axis=1)
+        
+        
+    def extrude(self, n_steps, desc=None):
+        # muy rapido
+        BUFFER = np.zeros((self.n_extrusion + n_steps, self.n_cross_section))
+        BUFFER[self.n_extrusion:] = self.VALUES
+        for buffer_index in tqdm(np.arange(n_steps)[::-1], desc=desc):
+            
+            new_values = self.A @ BUFFER[buffer_index + self.extrusion_sample_index + 1, self.cross_section_sample_index] + self.B @ np.random.standard_normal(size=self.B.shape[-1])
+            BUFFER[buffer_index] = new_values
+        self.VALUES = BUFFER[:self.n_extrusion]
+        
 
-        self.padding = (radius_sample_prop + beam_tol) * max_layer_beam_radii
-
-        for i_l, depth in enumerate(self.layer_depths):
-
-            MEAN_REL_POINTS = np.c_[self.MEAN_REL_X[i_l], self.MEAN_REL_Y[i_l]]
-            MEAN_REL_POINTS += 1e-12 * np.random.standard_normal(
-                size=MEAN_REL_POINTS.shape
-            )  # some jitter for the hull
-
-            hull = sp.spatial.ConvexHull(MEAN_REL_POINTS)
-            h_x, h_y = hull.points[hull.vertices].T
-            h_z = h_x + 1j * h_y
-            layer_hull_theta_z = h_z * (np.abs(h_z) + self.padding[i_l]) / np.abs(h_z)
-
-            self.layer_rotation_angles.append(
-                utils.get_minimal_bounding_rotation_angle(layer_hull_theta_z.ravel())
-            )
-
-            zop = (self.REL_X[i_l] + 1j * self.REL_Y[i_l]) * np.exp(1j * self.layer_rotation_angles[-1])
-            self.p[i_l], self.o[i_l] = np.real(zop), np.imag(zop)
-
-            res = self.min_ang_res[i_l].min()
-
-            para_ = np.arange(
-                self.p[i_l].min() - self.padding[i_l] - res,
-                self.p[i_l].max() + self.padding[i_l] + res,
-                res,
-            )
-            orth_ = np.arange(
-                self.o[i_l].min() - self.padding[i_l] - res,
-                self.o[i_l].max() + self.padding[i_l] + res,
-                res,
-            )
-
-            n_para, n_orth = len(para_), len(orth_)
-            self.lay_ang_res.append(res)
-
-            self.PARA_SPACING = np.gradient(para_).mean()
-            self.para.append(para_), self.orth.append(orth_)
-            self.n_para.append(len(para_)), self.n_orth.append(len(orth_))
-
-            ORTH_, PARA_ = np.meshgrid(orth_, para_)
-
-            self.genz.append(np.exp(-1j * self.layer_rotation_angles[-1]) * (PARA_[0] + 1j * ORTH_[0] - res))
-            XYZ = np.exp(-1j * self.layer_rotation_angles[-1]) * (PARA_ + 1j * ORTH_)
-
-            self.X.append(np.real(XYZ)), self.Y.append(np.imag(XYZ))
-            # self.O.append(ORTH_), self.P.append(PARA_)
-
-            del zop
-
-            para_i, orth_i = [], []
-            for ii, i in enumerate(
-                np.r_[
-                    0,
-                    2 ** np.arange(np.ceil(np.log(self.n_para[-1]) / np.log(2))),
-                    self.n_para[-1] - 1,
-                ]
-            ):
-
-                orth_i.append(
-                    np.unique(
-                        np.linspace(
-                            0,
-                            self.n_orth[-1] - 1,
-                            int(np.maximum(self.n_orth[-1] / (4**ii), 4)),
-                        ).astype(int)
-                    )
-                )
-                para_i.append(np.repeat(i, len(orth_i[-1])).astype(int))
-
-            self.AR_samples.append((np.concatenate(para_i), np.concatenate(orth_i)))
-
-            n_cm = len(self.AR_samples[-1][0])
-
-            if n_cm > 5000 and verbose:
-
-                warning_message = f"A very large covariance matrix for layer {i_l+1} (n_side = {n_cm})"
-                warnings.warn(warning_message)
-
-        self.C00, self.C01, self.C11, self.A, self.B = [], [], [], [], []
-
-        with tqdm(total=len(self.layer_depths), desc="Computing weights") as prog:
-            for i_l, (depth, LX, LY, AR, GZ) in enumerate(
-                zip(self.layer_depths, self.X, self.Y, self.AR_samples, self.genz)
-            ):
-
-                X0, Y0 = LX[AR], LY[AR]
-                X1, Y1 = np.real(GZ), np.imag(GZ)
-
-                R00 = np.sqrt(np.subtract.outer(X0, X0) ** 2 + np.subtract.outer(Y0, Y0) ** 2)
-                R01 = np.sqrt(np.subtract.outer(X1, X0) ** 2 + np.subtract.outer(Y1, Y0) ** 2)
-                R11 = np.sqrt(np.subtract.outer(X1, X1) ** 2 + np.subtract.outer(Y1, Y1) ** 2)
-
-                alpha = 1e-4
-
-                C00 = utils.approximate_matern(R00, self.outer_scale / depth, 5 / 6) + alpha * np.eye(
-                    len(X0)
-                )
-                C01 = utils.approximate_matern(R01, self.outer_scale / depth, 5 / 6)
-                C11 = utils.approximate_matern(R11, self.outer_scale / depth, 5 / 6) + alpha * np.eye(
-                    len(X1)
-                )
-
-                self.C00.append(C00)
-                self.C01.append(C01)
-                self.C11.append(C11)
-                self.A.append(np.matmul(C01, utils.fast_psd_inverse(C00)))
-                self.B.append(sp.linalg.lapack.dpotrf(C11 - np.matmul(self.A[-1], C01.T))[0])
-
-                prog.update(1)
-
-            if verbose:
-                print(
-                    "\n # | depth (m) | beam (m) | beam (') | sim (m) | sim (') | rms (mg/m2) | n_cov | orth | para | h2o (g/m3) | temp (K) | ws (m/s) | wb (deg) |"
-                )
-
-                for i_l, depth in enumerate(self.layer_depths):
-
-                    row_string  = f"{i_l+1:2} | {depth:9.01f} | {self.waists[i_l].min():8.02f} | {60*np.degrees(self.angular_waists[i_l].min()):8.02f} | "
-                    row_string += f"{depth*self.lay_ang_res[i_l]:7.02f} | {60*np.degrees(self.lay_ang_res[i_l]):7.02f} | "
-                    row_string += f"{1e3*self.layer_scaling[i_l].mean():11.02f} | {len(self.AR_samples[i_l][0]):5} | {self.n_orth[i_l]:4} | "
-                    row_string += f"{self.n_para[i_l]:4} | {1e3*self.water_vapor[i_l].mean():11.02f} | {self.temperature[i_l].mean():8.02f} | "
-                    row_string += f"{self.wind_speed[i_l].mean():8.02f} | {np.degrees(self.wind_bearing[i_l].mean()+np.pi):8.02f} |"
-                    print(row_string)
-
-    def atmosphere_timestep(self, i):  # iterate the i-th layer of atmosphere by one step
-
-        self.vals[i] = np.r_[
-            (
-                np.matmul(self.A[i], self.vals[i][self.AR_samples[i]])
-                + np.matmul(self.B[i], np.random.standard_normal(self.B[i].shape[0]))
-            )[None, :],
-            self.vals[i][:-1],
-        ]
-
-    def initialize_atmosphere(self, blurred=False):
-
-        self.vals = [np.zeros(lx.shape) for lx in self.X]
-        n_init_ = [n_para for n_para in self.n_para]
-        n_ts_ = [n_para for n_para in self.n_para]
-        tot_n_init, tot_n_ts = np.sum(n_init_), np.sum(n_ts_)
-        # self.gen_data = [np.zeros((n_ts,v.shape[1])) for n_ts,v in zip(n_ts_,self.lay_v_)]
-
-        with tqdm(total=tot_n_init, desc="Generating layers") as prog:
-            for i, n_init in enumerate(n_init_):
-                for i_init in range(n_init):
-
-                    self.atmosphere_timestep(i)
-                    prog.update(1)
-
-    def simulate_integrated_water_vapor(self, do_atmosphere=True, verbose=False):
+    def simulate_integrated_water_vapor(self):
 
         self.sim_start = ttime.time()
-        self.initialize_atmosphere()
+        self.extrude(n_steps=self.n_extrusion, desc="Generating atmosphere")
+        
+        #self.extrude(n_steps=0, desc="Initializing atmosphere")
 
-        self.rel_flucs = np.zeros(self.o.shape)
 
-        multichromatic_beams = False
+        trans_detector_angular_positions = (self.detector_offsets[:, None] + self.boresight_angular_position[None]) @ self.R.T
+        detector_values = np.zeros(trans_detector_angular_positions.shape[:-1])
+        for uband in self.array.ubands:
+            band_mask = self.array.dets.band == uband
+            angular_waist = self.angular_beam_waists[band_mask].mean()
+            F = self.array.make_filter(angular_waist, self.ang_res, self.array.beam_func)
+            FILTERED_VALUES = self.array.separably_filter(self.VALUES, F) 
+            detector_values[band_mask] = sp.interpolate.RegularGridInterpolator((self.cross_section_side, self.extrusion_side), 
+                                                                                FILTERED_VALUES.T)(trans_detector_angular_positions[band_mask])
 
-        with tqdm(total=self.n_layers, desc="Sampling layers") as prog:
+        # this is "zenith-scaled"
+        self.effective_integrated_water_vapor = 1. + self.site.pwv_rms * detector_values
 
-            for i_d, d in enumerate(self.layer_depths):
 
-                # Compute the filtered line-of-sight pwv corresponding to each layer
-
-                if multichromatic_beams:
-
-                    filtered_vals = np.zeros((len(self.array.nu), *self.vals[i_d].shape))
-                    angular_res = self.lay_ang_res[i_d]
-
-                    for i_f, f in enumerate(self.array.nu):
-
-                        angular_waist = self.angular_waists[i_d, i_f]
-
-                        self.F = self.array.make_filter(angular_waist, angular_res, self.array.beam_func)
-                        u, s, v = self.array.separate_filter(self.F)
-
-                        filtered_vals[i_f] = self.array.separably_filter(self.vals[i_d], u, s, v)
-
-                else:
-
-                    angular_res = self.lay_ang_res[i_d]
-                    angular_waist = self.angular_waists[i_d].mean()
-
-                    self.F = self.array.make_filter(angular_waist, angular_res, self.array.beam_func)
-
-                    filtered_vals = self.array.separably_filter(self.vals[i_d], self.F)
-                    self.rel_flucs[i_d] = sp.interpolate.RegularGridInterpolator((self.para[i_d], self.orth[i_d]), filtered_vals)(
-                        (self.p[i_d], self.o[i_d])
-                    )
-
-                prog.update(1)
-
-        # Convert PWV fluctuations to detector powers
-
-        self.effective_zenith_water_vapor = (self.rel_flucs * self.layer_scaling[:, None]).sum(axis=0)
-
-        self.effective_zenith_water_vapor *= 2e-2 * self.weather.column_water_vapor / self.effective_zenith_water_vapor.std()
-        self.effective_zenith_water_vapor += self.weather.column_water_vapor.mean()
-
-        self.integrated_water_vapor = self.effective_zenith_water_vapor / np.sin(self.EL)
-
-        # self.atm_power = np.zeros(self.effective_zenith_water_vapor.shape)
-
-        # with tqdm(total=len(self.array.ubands), desc='Integrating spectra') as prog:
-        #     for b in self.array.ubands:
-
-        #         bm = self.array.dets.bands == b
-
-        #         ba_am_trj = (self.am.t_rj * self.array.am_passbands[bm].mean(axis=0)[None,None,:]).sum(axis=-1)
-
-        #         BA_TRJ_RGI = sp.interpolate.RegularGridInterpolator((self.am.tcwv, np.radians(self.am.elev)), ba_am_trj)
-
-        #         self.atm_power[bm] = BA_TRJ_RGI((self.epwv[bm], self.EL[bm]))
-
-        #         prog.update(1)
-
-
-def get_extrusion_products(
-    time,
-    azim,
-    elev,
-    baseline,
-    field_of_view,
-    min_frequency,
-    wind_velocity,
-    wind_direction,
-    max_depth=3000,
-    extrusion_step=1,
-    min_res=10,
-    max_res=100,
-):
-    """
-    For the Kolomogorov-Taylor model. It works the same as one of those pasta extruding machines (https://youtu.be/TXtm_eNaIwQ). This function figures out the shape of the hole.
-
-    The inputs are:
-
-    azim, elev: time-ordered pointing
-    """
-
-    # Within this function, we work in the frame of wind-relative pointing phi'. The elevation stays the same. We define:
-    #
-    # phi' -> phi - phi_w
-    #
-    # where phi_w is the wind bearing, i.e. the direction the wind comes from. This means that:
-    #
-    # For phi' = 0°  you are looking into the wind
-    # For phi' = 90° the wind is going from left to right in your field of view.
-    #
-    # We further define the frame (x, y, z) = (r cos phi' sin theta, r sin phi' sin theta, r sin theta)
-
-    n_baseline = len(baseline)
-
-    # this converts the $(x, y, z)$ vertices of a beam looking straight up to the $(x, y, z)$ vertices of
-    # the time-ordered pointing of the beam in the frame of the wind direction.
-    elev_rotation_matrix = sp.spatial.transform.Rotation.from_euler("y", np.pi / 2 - elev).as_matrix()
-    azim_rotation_matrix = sp.spatial.transform.Rotation.from_euler(
-        "z", np.pi / 2 - azim + wind_direction
-    ).as_matrix()
-    total_rotation_matrix = np.matmul(azim_rotation_matrix, elev_rotation_matrix)
-
-    # this defines the maximum size of the beam that we have to worry about
-    max_beam_radius = _beam_sigma(max_depth, primary_size=primary_sizes, nu=5e9)
-    max_boresight_offset = max_beam_radius + max_depth * field_of_view / 2
-
-    # this returns a list of time-ordered vertices, which can used to construct a convex hull
-    time_ordered_vertices_list = []
-    for _min_radius, _max_radius, _baseline in zip(primary_sizes / 2, max_boresight_offset, baseline):
-
-        _rot_baseline = np.matmul(
-            sp.spatial.transform.Rotation.from_euler("z", wind_direction).as_matrix(),
-            _baseline,
-        )
-
-        _vertices = np.c_[
-            np.r_[[_.ravel() for _ in np.meshgrid([-_min_radius, _min_radius], [-_min_radius, _min_radius], [0])]],
-            np.r_[
-                [
-                    _.ravel()
-                    for _ in np.meshgrid(
-                        [-_max_radius, _max_radius],
-                        [-_max_radius, _max_radius],
-                        [max_depth],
-                    )
-                ]
-            ],
-        ]
-
-        time_ordered_vertices_list.append(
-            _rot_baseline[None] + np.swapaxes(np.matmul(total_rotation_matrix, _vertices), 1, -1).reshape(-1, 3)
-        )
-
-    # these are the bounds in space that we have to worry about. it has shape (3, 2)
-    extrusion_bounds = np.c_[
-        np.min([np.min(tov, axis=0) for tov in time_ordered_vertices_list], axis=0),
-        np.max([np.max(tov, axis=0) for tov in time_ordered_vertices_list], axis=0),
-    ]
-
-    # here we make the layers of the atmosphere given the bounds of $z$ and the specified resolution; we use
-    # regular layers to make interpolation more efficient later
-    min_height = np.max(
-        [np.max(tov[:, 2][tov[:, 2] < np.median(tov[:, 2])]) for tov in time_ordered_vertices_list]
-    )
-    max_height = np.min(
-        [np.min(tov[:, 2][tov[:, 2] > np.median(tov[:, 2])]) for tov in time_ordered_vertices_list]
-    )
-
-    height_samples = np.linspace(min_height, max_height, 1024)
-    dh = np.gradient(height_samples).mean()
-    dheight_dindex = np.interp(height_samples, [min_height, max_height], [min_res, max_res])
-    dindex_dheight = 1 / dheight_dindex
-    n_layers = int(np.sum(dindex_dheight * dh))
-    layer_heights = sp.interpolate.interp1d(
-        np.cumsum(dindex_dheight * dh),
-        height_samples,
-        bounds_error=False,
-        fill_value="extrapolate",
-    )(1 + np.arange(n_layers))
-
-    # here we define the cells through which turbulence will be extruded.
-    layer_res = np.gradient(layer_heights)
-    x_min, x_max = extrusion_bounds[0, 0], extrusion_bounds[0, 1]
-    n_per_layer = ((x_max - x_min) / layer_res).astype(int)
-    cell_x = np.concatenate(
-        [np.linspace(x_min, x_max, n) for i, (res, n) in enumerate(zip(layer_res, n_per_layer))]
-    )
-    cell_z = np.concatenate([h * np.ones(n) for i, (h, n) in enumerate(zip(layer_heights, n_per_layer))])
-    cell_res = np.concatenate([res * np.ones(n) for i, (res, n) in enumerate(zip(layer_res, n_per_layer))])
-
-    extrusion_cells = np.c_[cell_x, cell_z]
-    extrusion_shift = np.c_[cell_res, np.zeros(len(cell_z))]
-
-    eps = 1e-6
-
-    in_view = np.zeros(len(extrusion_cells)).astype(bool)
-    for tov in time_ordered_vertices_list:
-
-        baseline_in_view = np.zeros(len(extrusion_cells)).astype(bool)
-
-        hull = sp.spatial.ConvexHull(tov[:, [0, 2]])  # we only care about the $(x, z)$ dimensions here
-        A, b = hull.equations[:, :-1], hull.equations[:, -1:]
-
-        for shift_factor in [-1, 0, +1]:
-
-            baseline_in_view |= np.all(
-                (extrusion_cells + shift_factor * extrusion_shift) @ A.T + b.T < eps,
-                axis=1,
-            )
-
-        in_view |= baseline_in_view  # if the cell is in view of any of the baselines, keep it!
-
-        # plt.scatter(tov[:, 0], tov[:, 2], s=1e0)
-
-    # here we define the extrusion axis
-    y_min, y_max = (
-        extrusion_bounds[1, 0],
-        extrusion_bounds[1, 1] + wind_velocity * time.ptp(),
-    )
-    extrustion_axis = np.arange(y_min, y_max, extrusion_step)
-
-    return extrusion_cells[in_view].T, extrustion_axis, cell_res[in_view]
-
-
-class KolmogorovTaylorModel:
-    def __init__(
-        self,
-        time,
-        azim,
-        elev,
-        baseline,
-        field_of_view,
-        min_frequency,
-        wind_velocity,
-        wind_direction,
-        extrusion_step=1,
-        outer_scale=600,
-        min_res=10,
-        max_res=100,
-        max_depth=3000,
-    ):
-
-        self.time = time
-        self.azim = azim
-        self.elev = elev
-
-        self.baseline = baseline
-        self.field_of_view = field_of_view
-        self.min_frequency = min_frequency
-        self.wind_direction = wind_direction
-        self.wind_velocity = wind_velocity
-        self.extrusion_step = extrusion_step
-        self.min_res = min_res
-        self.max_res = max_res
-        self.max_depth = max_depth
-
-        self.outer_scale = outer_scale
-
-        (self.cX, self.cZ), self.Y, self.cres = get_extrusion_products(
-            self.time,
-            self.azim,
-            self.elev,
-            self.baseline,
-            self.field_of_view,
-            self.min_frequency,
-            self.wind_velocity,
-            self.wind_direction,
-            self.max_depth,
-            self.extrusion_step,
-            self.min_res,
-            self.max_res,
-        )
-
-        self.layer_heights, self.layer_index = np.unique(self.cZ, return_inverse=True)
-        self.n_layers = len(self.layer_heights)
-
-        self.dY, self.n_cells = np.gradient(self.Y).mean(), len(self.cX)
-        self.n_baseline = len(self.baseline)
-        self.n_extrusion = len(self.Y)
-        self.n_history = int(2 * self.outer_scale / self.dY) + 1
-
-        self.initialized = False
-
-    def initialize(self):
-
-        max_spacing = int(self.n_cells / 1.1)
-        iter_samples = [
-            np.random.choice(
-                self.n_cells,
-                int(self.n_cells / np.minimum(2**i, max_spacing)),
-                replace=False,
-            )
-            for i in range(self.n_history)
-        ]
-        self.hist_iter_index = np.concatenate(
-            [i * np.ones(len(index), dtype=int) for i, index in enumerate(iter_samples)]
-        )
-        self.cell_iter_index = np.concatenate(iter_samples).astype(int)
-
-        Xi, Xj = self.cX, self.cX[self.cell_iter_index]
-        Yi, Yj = np.zeros(self.n_cells), self.dY * (1 + self.hist_iter_index)
-        Zi, Zj = self.cZ, self.cZ[self.cell_iter_index]
-
-        Rii = np.sqrt(
-            np.subtract.outer(Xi, Xi) ** 2 + np.subtract.outer(Yi, Yi) ** 2 + np.subtract.outer(Zi, Zi) ** 2
-        )
-
-        Rij = np.sqrt(
-            np.subtract.outer(Xi, Xj) ** 2 + np.subtract.outer(Yi, Yj) ** 2 + np.subtract.outer(Zi, Zj) ** 2
-        )
-
-        Rjj = np.sqrt(
-            np.subtract.outer(Xj, Xj) ** 2 + np.subtract.outer(Yj, Yj) ** 2 + np.subtract.outer(Zj, Zj) ** 2
-        )
-
-        alpha = 1e-3
-
-        # this is all very computationally expensive stuff (n^3 is rough!)
-        self.Cii = utils.approximate_matern(
-            Rii, r0=self.outer_scale, nu=1 / 3, n_test_points=4096
-        ) + alpha**2 * np.eye(self.n_cells)
-        self.Cij = utils.approximate_matern(Rij, r0=self.outer_scale, nu=1 / 3, n_test_points=4096)
-        self.Cjj = utils.approximate_matern(Rjj, r0=self.outer_scale, nu=1 / 3, n_test_points=4096)
-        self.A = np.matmul(self.Cij, utils.fast_psd_inverse(self.Cjj))
-        self.B, _ = sp.linalg.lapack.dpotrf(self.Cii - np.matmul(self.A, self.Cij.T))
-
-        self.initialized = True
-
-    def extrude(self):
-
-        if not self.initialized:
-            self.initialize()
-
-        ARH = np.zeros((self.n_cells, self.n_history))
-        self.cdata = np.zeros((self.n_cells, self.n_extrusion))
-
-        def iterate_autoregression(ARH, n_iter):
-            for i in range(n_iter):
-                res = np.matmul(self.A, ARH[self.cell_iter_index, self.hist_iter_index]) + np.matmul(
-                    self.B, np.random.standard_normal(self.n_cells)
-                )
-                ARH = np.c_[res, ARH[:, :-1]]
-            return ARH
-
-        ARH = iterate_autoregression(ARH, n_iter=2 * self.n_history)
-
-        for i in range(self.n_extrusion):
-            ARH = iterate_autoregression(ARH, n_iter=1)
-            self.cdata[:, i] = ARH[:, 0]
