+++ maria/tests/test_atmosphere.py
@@ -1,12 +1,12 @@
 import pytest
 import maria
-from maria.atmosphere import LinearAngularSimulation
+from maria.atmosphere import SingleLayerSimulation
 
 
 @pytest.mark.atmosphere
 def test_linear_angular_model():
 
-    sim = LinearAngularSimulation(array="MUSTANG-2", pointing="daisy", site="GBT", atm_model="linear_angular")
+    sim = SingleLayerSimulation(array="MUSTANG-2", pointing="daisy", site="GBT")
     tod = sim.run()
